test: 로또 생성 및 비교 테스트
