+++ src/test/java/domain/MoneyTest.java
@@ -43,24 +43,4 @@ public class MoneyTest {
 
         Assertions.assertThat(money.calculateTotalLotto()).isEqualTo(14);
     }
-
-    @Test
-    @DisplayName("수익률 계산 테스트")
-    public void success_4() {
-        int input = 10000;
-        Money money = new Money(input);
-        Map<LottoRank, Integer> result = Map.of(
-            BOOM, 7,
-            FIFTH_PLACE, 1,
-            FOURTH_PLACE, 2,
-            THIRD_PLACE, 0,
-            SECOND_PLACE, 0,
-            FIRST_PLACE, 0
-        );
-
-        LottoStatistics lottoStatistics = new LottoStatistics(result);
-        int profit = lottoStatistics.calculateProfit();
-        double profitRate = money.calculateProfitRate(profit);
-        Assertions.assertThat(profitRate).isEqualTo(10.50);
-    }
 }
